+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - added support of ComponentContainer ([#515](https://github.com/deity-io/falcon/pull/515))
 - added auto-installing of Subscription handlers ([#520](https://github.com/deity-io/falcon/pull/520))
 - improved the way of extracting root fields for partial extension schemas ([#541](https://github.com/deity-io/falcon/pull/541))
+- fixed the usage of CacheProvider when working with non-scalar values ([#543](https://github.com/deity-io/falcon/pull/543))
 
 ### Falcon Client v0.5.3 (2019-08-01)